backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0

wget==3.2
pydantic>=2.10.0
python-dotenv==1.0.0
chromadb>=0.4.24,<0.6.0
boto3>=1.34.51,<1.35.0

numpy>=1.26.2,<2.0.0
pandas==2.0.3

sentence-transformers==2.6.1

langchain==0.3.26
langchain-core==0.3.70
langchain-community==0.3.26
langchain-aws>=0.2.0
langchain-groq==0.2.2
langchain-chroma==0.1.4

sqlalchemy>=2.0.25,<2.1.0

